Incremento y decremento añadidos

+++ mymath.h
@@ -1,18 +1,17 @@
 #ifndef MY_MATH_H
 #define MY_MATH_H
-
 /* mymath.h 
 	Contiene algunas funciones matemáticas importantes que se utilizarán
 	en el intérprete 
 */
 
-unsigned long long factorial(short n) {
+long double factorial(short n) {
     if(n <= 1)
         return 1;
-    unsigned long long r = 1;
+    long double r = 1.0;
     unsigned short i;
     for(i = 1; i <= n; i++)
-        r *= i;
+        r *= (long double)i;
     return r;
 }
